24 FEb

+++ BoilerPlate.cpp
@@ -3,6 +3,7 @@
 
 */
 
+
 #include <bits/stdc++.h>
 using namespace std;
 
@@ -10,7 +11,7 @@ int main(){
     ios::sync_with_stdio(false);
     cin.tie(NULL);
 
-    int tc;
+    long long tc;
     cin >> tc;
     while(tc--){
